+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.2.0
 
 * Added `animatedFitBounds` and `centerOnPoints` methods to `AnimatedMapController`
+* Properly disposed the `AnimationController` when the widget is disposed
+* Smoother animation when the `LatLng` is updated
 
 ## 0.1.0+1